pyproject.toml
[project]
name = "rag-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.70.0",
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.1",
    "supabase>=2.28.3",
]
